% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7.R
\name{add_role}
\alias{add_role}
\title{Add a role}
\usage{
add_role(client, ...)
}
\arguments{
\item{client}{An IxAccessClient object}

\item{role}{The role name to add}
}
\value{
NULL (invisibly)
}
\description{
Add a role
}
