% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7.R
\name{add_roles}
\alias{add_roles}
\title{Add multiple roles}
\usage{
add_roles(client, ...)
}
\arguments{
\item{client}{An IxAccessClient object}

\item{roles}{A character vector of role names to add}
}
\value{
NULL (invisibly)
}
\description{
Add multiple roles
}
