% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7.R
\name{assign_resource_to_role}
\alias{assign_resource_to_role}
\title{Assign a resource to a role}
\usage{
assign_resource_to_role(client, ...)
}
\arguments{
\item{client}{An IxAccessClient object}

\item{role}{The role name}

\item{resource_tag}{The resource tag}

\item{resource_value}{The resource value}
}
\value{
NULL (invisibly)
}
\description{
Assign a resource to a role
}
