% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7.R
\name{assign_role}
\alias{assign_role}
\title{Assign a role to another role}
\usage{
assign_role(client, ...)
}
\arguments{
\item{client}{An IxAccessClient object}

\item{assignee}{The role receiving the assignment}

\item{role}{The role being assigned}
}
\value{
NULL (invisibly)
}
\description{
Assign a role to another role
}
