% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7.R
\name{delete_role}
\alias{delete_role}
\title{Delete a role}
\usage{
delete_role(client, ...)
}
\arguments{
\item{client}{An IxAccessClient object}

\item{role}{The role name to delete}
}
\value{
NULL (invisibly)
}
\description{
Delete a role
}
