% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7.R
\name{exists_role}
\alias{exists_role}
\title{Check if a role exists}
\usage{
exists_role(client, ...)
}
\arguments{
\item{client}{An IxAccessClient object}

\item{role}{The role name}
}
\value{
A logical value indicating if the role exists
}
\description{
Check if a role exists
}
