% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7.R
\name{find_roles_with_resource}
\alias{find_roles_with_resource}
\title{Find roles with a specific resource}
\usage{
find_roles_with_resource(client, ...)
}
\arguments{
\item{client}{An IxAccessClient object}

\item{tag}{The resource tag}

\item{value}{The resource value}
}
\value{
A character vector of role names
}
\description{
Find roles with a specific resource
}
