% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7.R
\name{find_roles_with_resource_tag}
\alias{find_roles_with_resource_tag}
\title{Find roles with a specific resource tag}
\usage{
find_roles_with_resource_tag(client, ...)
}
\arguments{
\item{client}{An IxAccessClient object}

\item{tag}{The resource tag}
}
\value{
A character vector of role names
}
\description{
Find roles with a specific resource tag
}
