% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7.R
\name{get_all_resources_for_role}
\alias{get_all_resources_for_role}
\title{Get all resources for a role}
\usage{
get_all_resources_for_role(client, ...)
}
\arguments{
\item{client}{An IxAccessClient object}

\item{role}{The role name}
}
\value{
A list of character vectors, where each element is a resource tag and its value is a vector of resource values
}
\description{
Get all resources for a role
}
