% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7.R
\name{get_all_resources_for_role_by_tag}
\alias{get_all_resources_for_role_by_tag}
\title{Get all resources for a role by tag}
\usage{
get_all_resources_for_role_by_tag(client, ...)
}
\arguments{
\item{client}{An IxAccessClient object}

\item{role}{The role name}

\item{tag}{The resource tag}
}
\value{
A character vector of resource values
}
\description{
Get all resources for a role by tag
}
