% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7.R
\name{has_resource}
\alias{has_resource}
\title{Check if a role has access to a specific resource}
\usage{
has_resource(client, ...)
}
\arguments{
\item{client}{An IxAccessClient object}

\item{role}{The role name}

\item{resource_tag}{The resource tag}

\item{resource_value}{The resource value}
}
\value{
A logical value indicating if the role has access to the resource
}
\description{
Check if a role has access to a specific resource
}
