% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7.R
\name{has_role}
\alias{has_role}
\title{Check if an assignee has a specific role}
\usage{
has_role(client, ...)
}
\arguments{
\item{client}{An IxAccessClient object}

\item{assignee}{The assignee role name}

\item{role}{The role name to check for}
}
\value{
A logical value indicating if the assignee has the role
}
\description{
Check if an assignee has a specific role
}
