% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7.R
\name{list_all_roles_for_role}
\alias{list_all_roles_for_role}
\title{List all roles for a given role}
\usage{
list_all_roles_for_role(client, ...)
}
\arguments{
\item{client}{An IxAccessClient object}

\item{role}{The role name}
}
\value{
A character vector of role names
}
\description{
List all roles for a given role
}
