% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7.R
\name{list_members_of}
\alias{list_members_of}
\title{List members of a role}
\usage{
list_members_of(client, ...)
}
\arguments{
\item{client}{An IxAccessClient object}

\item{role}{The role name}
}
\value{
A character vector of role names
}
\description{
List members of a role
}
