% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7.R
\name{list_roles}
\alias{list_roles}
\title{List all roles}
\usage{
list_roles(client, ...)
}
\arguments{
\item{client}{An IxAccessClient object}
}
\value{
A character vector of role names
}
\description{
List all roles
}
