% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7.R
\name{unassign_resource_from_role}
\alias{unassign_resource_from_role}
\title{Unassign a resource from a role}
\usage{
unassign_resource_from_role(client, ...)
}
\arguments{
\item{client}{An IxAccessClient object}

\item{role}{The role name}

\item{resource_tag}{The resource tag}

\item{resource_value}{The resource value}
}
\value{
NULL (invisibly)
}
\description{
Unassign a resource from a role
}
