% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7.R
\name{unassign_role}
\alias{unassign_role}
\title{Unassign a role from another role}
\usage{
unassign_role(client, ...)
}
\arguments{
\item{client}{An IxAccessClient object}

\item{assignee}{The role to remove the assignment from}

\item{role}{The role being unassigned}
}
\value{
NULL (invisibly)
}
\description{
Unassign a role from another role
}
